+++ domain/src/main/java/com/tboy/domain/model/CurrentWeather.kt
@@ -4,12 +4,12 @@ data class CurrentWeather(
     val time: Int? = null,
     val summary: String? = null,
     val icon: String? = null,
-    val precipIntensity: Float? = null,
-    val precipProbability: Float? = null,
-    val temperature: Float? = null,
-    val humidity: Float? = null,
-    val apparentTemperature: Float? = null,
-    val pressure: Float? = null,
-    val windSpeed: Float? = null,
-    val cloudCover: Float? = null
+    val precipIntensity: Double? = null,
+    val precipProbability: Double? = null,
+    val temperature: Double? = null,
+    val humidity: Double? = null,
+    val apparentTemperature: Double? = null,
+    val pressure: Double? = null,
+    val windSpeed: Double? = null,
+    val cloudCover: Double? = null
 )
